+++ examples/swift/Shared/testingApp.swift
@@ -12,7 +12,7 @@ import SwiftUI
 struct testingApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            NavigationViews()
         }
     }
 }