+++ examples/vue/nuxt-3/app.vue
@@ -20,6 +20,7 @@
 
 <script setup>
 import { RenderContent, getContent, isPreviewing } from '@builder.io/sdk-vue/vue3';
+import '@builder.io/sdk-vue/vue3/css';
 
 import HelloWorldComponent from './components/HelloWorld.vue';