+++ packages/sdks-tests/src/specs/editing-styles.ts
@@ -1,0 +1,161 @@
+export const EDITING_STYLES = {
+  lastUpdatedBy: 'OcOewqA7uqVVlVfqY453F8vgcc33',
+  folders: [],
+  data: {
+    newField3: 'testing',
+    themeId: false,
+    title: 'editing-styles',
+    blocks: [
+      {
+        '@type': '@builder.io/sdk:Element',
+        '@version': 2,
+        id: 'builder-d90e0e526f9a4ec0839b70774f2340b5',
+        component: {
+          name: 'Core:Button',
+          options: {
+            text: 'Click me!',
+            openLinkInNewTab: false,
+            link: '/b',
+          },
+        },
+        responsiveStyles: {
+          large: {
+            display: 'flex',
+            flexDirection: 'column',
+            position: 'relative',
+            flexShrink: '0',
+            boxSizing: 'border-box',
+            marginTop: '20px',
+            appearance: 'none',
+            paddingTop: '15px',
+            paddingBottom: '15px',
+            paddingLeft: '25px',
+            paddingRight: '25px',
+            backgroundColor: 'rgb(184, 35, 35)',
+            color: 'white',
+            borderRadius: '4px',
+            textAlign: 'center',
+            cursor: 'pointer',
+          },
+        },
+      },
+    ],
+    url: '/editing-styles',
+    state: {
+      deviceSize: 'large',
+      location: {
+        pathname: '/editing-styles',
+        path: ['editing-styles'],
+        query: {},
+      },
+    },
+  },
+  modelId: '240a12053d674735ac2a384dcdc561b5',
+  query: [
+    {
+      '@type': '@builder.io/core:Query',
+      property: 'urlPath',
+      value: '/editing-styles',
+      operator: 'is',
+    },
+  ],
+  published: 'published',
+  screenshot:
+    'https://cdn.builder.io/api/v1/image/assets%2Ff1a790f8c3204b3b8c5c1795aeac4660%2F02f1a5144a4f417988a33c5025ea854b',
+  firstPublished: 1698086079699,
+  testRatio: 1,
+  lastUpdated: 1698088978247,
+  createdDate: 1698085958541,
+  createdBy: 'OcOewqA7uqVVlVfqY453F8vgcc33',
+  variations: {},
+  meta: {
+    lastPreviewUrl:
+      'https://svelte-vite-example.vercel.app/editing-styles?builder.space=f1a790f8c3204b3b8c5c1795aeac4660&builder.cachebust=true&builder.preview=page&builder.noCache=true&builder.allowTextEdit=true&__builder_editing__=true&builder.overrides.page=7afe9144c71d459ab54af49a16207179&builder.overrides.7afe9144c71d459ab54af49a16207179=7afe9144c71d459ab54af49a16207179&builder.overrides.page:/editing-styles=7afe9144c71d459ab54af49a16207179&builder.options.locale=Default',
+    kind: 'page',
+    hasLinks: false,
+  },
+  name: 'editing-styles',
+  id: '7afe9144c71d459ab54af49a16207179',
+  rev: 'opsj9u54avk',
+};
+
+export const MODIFIED_EDITING_STYLES = {
+  lastUpdatedBy: 'OcOewqA7uqVVlVfqY453F8vgcc33',
+  folders: [],
+  data: {
+    newField3: 'testing',
+    themeId: false,
+    title: 'editing-styles',
+    blocks: [
+      {
+        '@type': '@builder.io/sdk:Element',
+        '@version': 2,
+        id: 'builder-d90e0e526f9a4ec0839b70774f2340b5',
+        component: {
+          name: 'Core:Button',
+          options: {
+            text: 'Click me!',
+            openLinkInNewTab: false,
+            link: '/b',
+          },
+        },
+        responsiveStyles: {
+          large: {
+            display: 'flex',
+            flexDirection: 'column',
+            position: 'relative',
+            flexShrink: '0',
+            boxSizing: 'border-box',
+            marginTop: '20px',
+            appearance: 'none',
+            paddingTop: '15px',
+            paddingBottom: '15px',
+            paddingLeft: '25px',
+            paddingRight: '25px',
+            backgroundColor: 'rgb(19, 67, 92)',
+            color: 'white',
+            borderRadius: '4px',
+            textAlign: 'center',
+            cursor: 'pointer',
+          },
+        },
+      },
+    ],
+    url: '/editing-styles',
+    state: {
+      deviceSize: 'large',
+      location: {
+        pathname: '/editing-styles',
+        path: ['editing-styles'],
+        query: {},
+      },
+    },
+  },
+  modelId: '240a12053d674735ac2a384dcdc561b5',
+  query: [
+    {
+      '@type': '@builder.io/core:Query',
+      property: 'urlPath',
+      value: '/editing-styles',
+      operator: 'is',
+    },
+  ],
+  published: 'published',
+  screenshot:
+    'https://cdn.builder.io/api/v1/image/assets%2Ff1a790f8c3204b3b8c5c1795aeac4660%2F02f1a5144a4f417988a33c5025ea854b',
+  firstPublished: 1698086079699,
+  testRatio: 1,
+  lastUpdated: 1698088978247,
+  createdDate: 1698085958541,
+  createdBy: 'OcOewqA7uqVVlVfqY453F8vgcc33',
+  variations: {},
+  meta: {
+    lastPreviewUrl:
+      'https://svelte-vite-example.vercel.app/editing-styles?builder.space=f1a790f8c3204b3b8c5c1795aeac4660&builder.cachebust=true&builder.preview=page&builder.noCache=true&builder.allowTextEdit=true&__builder_editing__=true&builder.overrides.page=7afe9144c71d459ab54af49a16207179&builder.overrides.7afe9144c71d459ab54af49a16207179=7afe9144c71d459ab54af49a16207179&builder.overrides.page:/editing-styles=7afe9144c71d459ab54af49a16207179&builder.options.locale=Default',
+    kind: 'page',
+    hasLinks: false,
+  },
+  name: 'editing-styles',
+  id: '7afe9144c71d459ab54af49a16207179',
+  rev: 'opsj9u54avk',
+};