+++ packages/sdks-tests/src/specs/index.ts
@@ -21,7 +21,9 @@ import { CONTENT as reactiveState } from './reactive-state.js';
 import { CONTENT as showHideIf } from './show-hide-if.js';
 import { CONTENT as textBlock } from './text-block.js';
 import { CONTENT as stateBinding } from './state-binding.js';
+import { CONTENT as nestedSymbols } from './nested-symbols.js';
 import type { BuilderContent } from './types.js';
+import { EDITING_STYLES } from './editing-styles.js';
 
 function isBrowser(): boolean {
   return typeof window !== 'undefined' && typeof document !== 'undefined';
@@ -57,6 +59,8 @@ const pages = {
   '/custom-breakpoints-reset': customBreakpointsReset,
   '/text-block': textBlock,
   '/state-binding': stateBinding,
+  '/nested-symbols': nestedSymbols,
+  '/editing-styles': EDITING_STYLES,
 } as const;
 
 const apiVersionPathToProp = {