+++ packages/sdks/src/blocks/text/text.lite.tsx
@@ -1,6 +1,6 @@
 export default function Text(props: { text?: string }) {
   return (
-    <span
+    <div
       class={
         /* NOTE: This class name must be "builder-text" for inline editing to work in the Builder editor */
         'builder-text'