+++ packages/sdks/src/components/content/components/enable-editor.lite.tsx
@@ -17,24 +17,25 @@ import {
   useRef,
   setContext,
   useTarget,
+  onEvent,
 } from '@builder.io/mitosis';
 import {
   registerInsertMenu,
   setupBrowserForEditing,
 } from '../../../scripts/init-editing.js';
-import { checkIsDefined } from '../../../helpers/nullable.js';
 import { getInteractionPropertiesForEvent } from '../../../functions/track/interaction.js';
 import type {
   ContentProps,
   BuilderComponentStateChange,
 } from '../content.types.js';
-import { logger } from '../../../helpers/logger.js';
 import type { ComponentInfo } from '../../../types/components.js';
 import { fetchOneEntry } from '../../../functions/get-content/index.js';
 import { isPreviewing } from '../../../functions/is-previewing.js';
 import type { BuilderContent } from '../../../types/builder-content.js';
 import { postPreviewContent } from '../../../helpers/preview-lru-cache/set.js';
 import { fastClone } from '../../../functions/fast-clone.js';
+import { logger } from '../../../helpers/logger.js';
+import { getDefaultCanTrack } from '../../../helpers/canTrack.js';
 
 useMetadata({
   qwik: {
@@ -57,9 +58,11 @@ type BuilderEditorProps = Omit<
 };
 
 export default function EnableEditor(props: BuilderEditorProps) {
+  /**
+   * This var name is hard-coded in some Mitosis Plugins. Do not change.
+   */
   const elementRef = useRef<HTMLDivElement>();
   const state = useStore({
-    canTrackToUse: checkIsDefined(props.canTrack) ? props.canTrack : true,
     forceReRenderCount: 0,
     mergeNewContent(newContent: BuilderContent) {
       const newContentValue = {
@@ -159,7 +162,7 @@ export default function EnableEditor(props: BuilderEditorProps) {
         const contentId = props.builderContextSignal.value.content?.id;
         _track({
           type: 'click',
-          canTrack: state.canTrackToUse,
+          canTrack: getDefaultCanTrack(props.canTrack),
           contentId,
           apiKey: props.apiKey,
           variationId: variationId !== contentId ? variationId : undefined,
@@ -263,94 +266,146 @@ export default function EnableEditor(props: BuilderEditorProps) {
     }
   });
 
-  onMount(() => {
-    if (!props.apiKey) {
-      logger.error(
-        'No API key provided to `RenderContent` component. This can cause issues. Please provide an API key using the `apiKey` prop.'
+  onEvent(
+    'initeditingbldr',
+    () => {
+      state.forceReRenderCount = state.forceReRenderCount + 1;
+      window.addEventListener('message', state.processMessage);
+
+      registerInsertMenu();
+      setupBrowserForEditing({
+        ...(props.locale ? { locale: props.locale } : {}),
+        ...(props.includeRefs ? { includeRefs: props.includeRefs } : {}),
+        ...(props.enrich ? { enrich: props.enrich } : {}),
+      });
+      Object.values<ComponentInfo>(
+        props.builderContextSignal.value.componentInfos
+      ).forEach((registeredComponent) => {
+        const message = createRegisterComponentMessage(registeredComponent);
+        window.parent?.postMessage(message, '*');
+      });
+      window.addEventListener(
+        'builder:component:stateChangeListenerActivated',
+        state.emitStateUpdate
       );
-    }
+    },
+    elementRef,
+    true
+  );
 
-    if (isBrowser()) {
-      if (isEditing()) {
-        state.forceReRenderCount = state.forceReRenderCount + 1;
-        window.addEventListener('message', state.processMessage);
+  onEvent(
+    'initpreviewingbldr',
+    () => {
+      const searchParams = new URL(location.href).searchParams;
+      const searchParamPreviewModel = searchParams.get('builder.preview');
+      const searchParamPreviewId = searchParams.get(
+        `builder.preview.${searchParamPreviewModel}`
+      );
+      const previewApiKey =
+        searchParams.get('apiKey') || searchParams.get('builder.space');
 
-        registerInsertMenu();
-        setupBrowserForEditing({
-          ...(props.locale ? { locale: props.locale } : {}),
-          ...(props.includeRefs ? { includeRefs: props.includeRefs } : {}),
-          ...(props.enrich ? { enrich: props.enrich } : {}),
+      /**
+       * Make sure that:
+       * - the preview model name is the same as the one we're rendering, since there can be multiple models rendered
+       *  at the same time, e.g. header/page/footer.
+       * - the API key is the same, since we don't want to preview content from other organizations.
+       * - if there is content, that the preview ID is the same as that of the one we receive.
+       *
+       * TO-DO: should we only update the state when there is a change?
+       **/
+      if (
+        searchParamPreviewModel === props.model &&
+        previewApiKey === props.apiKey &&
+        (!props.content || searchParamPreviewId === props.content.id)
+      ) {
+        fetchOneEntry({
+          model: props.model,
+          apiKey: props.apiKey,
+          apiVersion: props.builderContextSignal.value.apiVersion,
+        }).then((content) => {
+          if (content) {
+            state.mergeNewContent(content);
+          }
         });
-        Object.values<ComponentInfo>(
-          props.builderContextSignal.value.componentInfos
-        ).forEach((registeredComponent) => {
-          const message = createRegisterComponentMessage(registeredComponent);
-          window.parent?.postMessage(message, '*');
-        });
-        window.addEventListener(
-          'builder:component:stateChangeListenerActivated',
-          state.emitStateUpdate
-        );
       }
-      if (props.builderContextSignal.value.content) {
-        const variationId =
-          props.builderContextSignal.value.content?.testVariationId;
-        const contentId = props.builderContextSignal.value.content?.id;
+    },
+    elementRef,
+    true
+  );
+
+  onMount(() => {
+    if (isBrowser()) {
+      if (isEditing() && elementRef) {
+        useTarget({
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+          // @ts-ignore
+          solid: () => INJECT_EDITING_HOOK_HERE,
+          default: () => {
+            elementRef.dispatchEvent(new CustomEvent('initeditingbldr'));
+          },
+        });
+      }
+
+      const shouldTrackImpression = useTarget({
+        qwik:
+          elementRef.attributes.getNamedItem('shouldTrack')?.value === 'true',
+        default:
+          props.builderContextSignal.value.content &&
+          getDefaultCanTrack(props.canTrack),
+      });
+
+      if (shouldTrackImpression) {
+        const variationId = useTarget({
+          qwik: elementRef.attributes.getNamedItem('variationId')?.value,
+          default: props.builderContextSignal.value.content?.testVariationId,
+        });
+        const contentId = useTarget({
+          qwik: elementRef.attributes.getNamedItem('contentId')?.value,
+          default: props.builderContextSignal.value.content?.id,
+        });
+        const apiKeyProp = useTarget({
+          qwik: elementRef.attributes.getNamedItem('apiKey')?.value,
+          default: props.apiKey,
+        });
+
         _track({
           type: 'impression',
-          canTrack: state.canTrackToUse,
+          canTrack: true,
           contentId,
-          apiKey: props.apiKey,
+          apiKey: apiKeyProp!,
           variationId: variationId !== contentId ? variationId : undefined,
         });
       }
 
-      useTarget({
-        rsc: () => {},
-        default: () => {
-          // override normal content in preview mode
-          if (isPreviewing()) {
-            const searchParams = new URL(location.href).searchParams;
-            const searchParamPreviewModel = searchParams.get('builder.preview');
-            const searchParamPreviewId = searchParams.get(
-              `builder.preview.${searchParamPreviewModel}`
-            );
-            const previewApiKey =
-              searchParams.get('apiKey') || searchParams.get('builder.space');
+      // override normal content in preview mode
+      if (isPreviewing() && elementRef) {
+        useTarget({
+          rsc: () => {},
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+          // @ts-ignore
+          solid: () => INJECT_PREVIEWING_HOOK_HERE,
+          default: () => {
+            elementRef.dispatchEvent(new CustomEvent('initpreviewingbldr'));
+          },
+        });
+      }
+    }
+  });
 
-            /**
-             * Make sure that:
-             * - the preview model name is the same as the one we're rendering, since there can be multiple models rendered
-             *  at the same time, e.g. header/page/footer.
-             * - the API key is the same, since we don't want to preview content from other organizations.
-             * - if there is content, that the preview ID is the same as that of the one we receive.
-             *
-             * TO-DO: should we only update the state when there is a change?
-             **/
-            if (
-              searchParamPreviewModel === props.model &&
-              previewApiKey === props.apiKey &&
-              (!props.content || searchParamPreviewId === props.content.id)
-            ) {
-              fetchOneEntry({
-                model: props.model,
-                apiKey: props.apiKey,
-                apiVersion: props.builderContextSignal.value.apiVersion,
-              }).then((content) => {
-                if (content) {
-                  state.mergeNewContent(content);
-                }
-              });
-            }
-          }
-        },
-      });
+  onMount(
+    () => {
+      if (!props.apiKey) {
+        logger.error(
+          'No API key provided to `RenderContent` component. This can cause issues. Please provide an API key using the `apiKey` prop.'
+        );
+      }
 
       state.evaluateJsCode();
       state.runHttpRequests();
       state.emitStateUpdate();
-    }
-  });
+    },
+    { onSSR: true }
+  );
 
   onUpdate(() => {
     state.evaluateJsCode();
@@ -370,6 +425,19 @@ export default function EnableEditor(props: BuilderEditorProps) {
   return (
     <Show when={props.builderContextSignal.value.content}>
       <div
+        {...useTarget({
+          qwik: {
+            apiKey: props.apiKey,
+            contentId: props.builderContextSignal.value.content?.id,
+            variationId:
+              props.builderContextSignal.value.content?.testVariationId,
+            shouldTrack: String(
+              props.builderContextSignal.value.content &&
+                getDefaultCanTrack(props.canTrack)
+            ),
+          },
+          default: {},
+        })}
         key={state.forceReRenderCount}
         ref={elementRef}
         onClick={(event) => state.onClick(event)}