+++ packages/sdks/src/functions/fetch-builder-props.ts
@@ -43,7 +43,21 @@ type GetBuilderPropsOptions = (Omit<GetContentOptions, 'model'> & {
   );
 
 /**
- * Fetches builder content, and returns it along with sensible defaults for other props that `Content` needs to render.
+ * Given an `apiKey` and `url` (or `path` + `searchParams`), provides all props that `Content` needs to render Builder Content.
+ *
+ * @example
+ * ```jsx
+ * const builderProps = await fetchBuilderProps({
+ *    apiKey: 'API_KEY',
+ *    // provide `url`
+ *    url: yourPageUrl,
+ *    // OR provide `path` + `searchParams`
+ *    path: yourPath,
+ *    searchParams: yourSearchParams,
+ * });
+ *
+ * return <Content {...builderProps} />;
+ * ```
  */
 export const fetchBuilderProps = async (
   _args: GetBuilderPropsOptions