+++ packages/core/CHANGELOG.md
@@ -1,3 +1,15 @@
+### 2.1.1
+
+- Accept `authToken` option in `getAll` to fetch private models.
+
+### 2.1.0
+
+- Setting `noTraverse` option's default to `true` when fetching multiple content entries.
+
+### 2.0.8
+
+- Fix: remove `setImmediate` usage to fix issue with next.js edge runtime.
+
 ### 2.0.7
 
 - Fix: remove dev dependency from `dependencies` (`nx` and `nx-cloud`)